Organizando o codigo em classes
