+++ src/main/java/br/com/webscraping/Scraping/App.java
@@ -1,45 +1,21 @@
 package br.com.webscraping.Scraping;
 
-import java.io.FileInputStream;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSession;
-import javax.net.ssl.X509TrustManager;
-
-import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
+import br.com.webscraping.entity.Limite;
+import br.com.webscraping.entity.Vertice;
+import br.com.webscraping.jsoup.Connect;
+import br.com.webscraping.report.CriarExcel;
+import br.com.webscraping.util.Util;
+
 public class App {
-	private static final String USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36(KHTML, like Gecko) Chrome/80.0.3987.163 Safari/537.36";
 
 	public static void main(String[] args) throws Exception {
-		/*
-		 * GET START
-		 * 
-		 * Document document = Jsoup.connect("https://www.imdb.com/chart/top").get();
-		 * 
-		 * for (Element row : document.select("table.chart.full-width tr")) {
-		 * 
-		 * String title = row.select(".titleColumn a").text(); // String title =
-		 * row.select(".titleColumn").text(); String rating =
-		 * row.select(".imdbRating").text();
-		 * 
-		 * System.out.println(title + "->Rating: " + rating); }
-		 */
-
-		// SIGEF
 
 		var paginaAtual = 1;
 		Boolean ultimaPaginaEncontrada = false;
@@ -55,16 +31,14 @@ public class App {
 			System.out.println("Pagina atual: " + paginaAtual);
 
 			String url = "https://sigef.incra.gov.br/consultar/parcelas/?page=%s&termo=&pesquisa_avancada=True&cpf_cnpj=&proprietario=&cns=%s&matricula=&codigo=&protocolo=&credenciado=&vertice=&sncr=";
-			String urlFormatada = String.format(url, paginaAtual, cns);
+			String urlConsulta = String.format(url, paginaAtual, cns);
 
-			enableSSLSocket();
-			Document document = Jsoup.connect(urlFormatada).timeout(0).userAgent(USER_AGENT).data("name", "jsoup")
-					.get();
+			Connect.enableSSLSocket();
+			Document document = Connect.jsoupConnect(urlConsulta);
 
-		
 			if (ultimaPaginaEncontrada == false) {
 
-				//localizar a numero da ultima pagina
+				// localizar a numero da ultima pagina
 				Element paginacao = document.getElementsByClass("pagination").first();
 
 				Elements lis = paginacao.selectFirst("ul").getElementsByTag("li");
@@ -94,17 +68,16 @@ public class App {
 				try {
 
 					System.out.println("Parcela id " + id);
-					enableSSLSocket();
-					detalhesDoc = Jsoup.connect("https://sigef.incra.gov.br/geo/parcela/detalhe/" + id + "/").timeout(0)
-							.userAgent(USER_AGENT).data("name", "jsoup").get();
+					String urlParcela = "https://sigef.incra.gov.br/geo/parcela/detalhe/" + id + "/";
+					Connect.enableSSLSocket();
+					detalhesDoc = Connect.jsoupConnect(urlParcela);
 
 				} catch (Exception e) {
 					var xx = e.getMessage();
 					System.out.println(xx);
 				}
-			
 
-				Element verticesTable = encontreTabelaPorNome(detalhesDoc, "Vértices da parcela");
+				Element verticesTable = Util.encontreTabelaPorNome(detalhesDoc, "Vértices da parcela");
 				Elements linhasVertices = verticesTable.getElementsByTag("tr");
 
 				int totalVertices = linhasVertices.size();
@@ -128,7 +101,7 @@ public class App {
 					vertices.add(vertice);
 				}
 
-				Element limitesTable = encontreTabelaPorNome(detalhesDoc, "Limites");
+				Element limitesTable = Util.encontreTabelaPorNome(detalhesDoc, "Limites");
 				Elements linhaslimites = limitesTable.getElementsByTag("tr");
 
 				int totalLimites = linhaslimites.size();
@@ -162,32 +135,4 @@ public class App {
 
 	}
 
-	private static Element encontreTabelaPorNome(Document document, String cabecaho) {
-		var divParcelasCabecalho = document.selectFirst("div:containsOwn(" + cabecaho +")");
-		var divParcelasConteudo = divParcelasCabecalho.nextElementSibling();
-		return divParcelasConteudo.selectFirst("table");
-	}
-
-	// https://nanashi07.blogspot.com/2014/06/enable-ssl-connection-for-jsoup.html
-	public static void enableSSLSocket() throws KeyManagementException, NoSuchAlgorithmException {
-		HttpsURLConnection.setDefaultHostnameVerifier(new HostnameVerifier() {
-			public boolean verify(String hostname, SSLSession session) {
-				return true;
-			}
-		});
-
-		SSLContext context = SSLContext.getInstance("TLS");
-		context.init(null, new X509TrustManager[] { new X509TrustManager() {
-			public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-			}
-
-			public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
-			}
-
-			public X509Certificate[] getAcceptedIssuers() {
-				return new X509Certificate[0];
-			}
-		} }, new SecureRandom());
-		HttpsURLConnection.setDefaultSSLSocketFactory(context.getSocketFactory());
-	}
 }
