--- src/main/java/br/com/webscraping/Scraping/Util.java
@@ -1,3 +1,0 @@
-package br.com.webscraping.Scraping;
-
-
