+++ src/main/java/br/com/webscraping/jsoup/Connect.java
@@ -1,0 +1,50 @@
+package br.com.webscraping.jsoup;
+
+import java.io.IOException;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.HttpsURLConnection;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSession;
+import javax.net.ssl.X509TrustManager;
+
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+
+public class Connect {
+
+	// https://nanashi07.blogspot.com/2014/06/enable-ssl-connection-for-jsoup.html
+	public static void enableSSLSocket() throws KeyManagementException, NoSuchAlgorithmException {
+		HttpsURLConnection.setDefaultHostnameVerifier(new HostnameVerifier() {
+			public boolean verify(String hostname, SSLSession session) {
+				return true;
+			}
+		});
+
+		SSLContext context = SSLContext.getInstance("TLS");
+		context.init(null, new X509TrustManager[] { new X509TrustManager() {
+			public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+			}
+
+			public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+			}
+
+			public X509Certificate[] getAcceptedIssuers() {
+				return new X509Certificate[0];
+			}
+		} }, new SecureRandom());
+		HttpsURLConnection.setDefaultSSLSocketFactory(context.getSocketFactory());
+	}
+
+	public static Document jsoupConnect(String urlFormatada) throws IOException {
+		String USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36(KHTML, like Gecko) Chrome/80.0.3987.163 Safari/537.36";
+		return Jsoup.connect(urlFormatada).timeout(0).userAgent(USER_AGENT).data("name", "jsoup").get();
+
+	}
+
+}
