+++ src/main/java/br/com/webscraping/util/Util.java
@@ -1,0 +1,14 @@
+package br.com.webscraping.util;
+
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
+public class Util {
+
+	public static Element encontreTabelaPorNome(Document document, String cabecaho) {
+		var divParcelasCabecalho = document.selectFirst("div:containsOwn(" + cabecaho + ")");
+		var divParcelasConteudo = divParcelasCabecalho.nextElementSibling();
+		return divParcelasConteudo.selectFirst("table");
+	}
+
+}
